Add files via upload
Specialized PCB for precise control of a laser system, allowing adjustment of both power and operating frequency through a user-friendly interface. It features an LCD screen that displays  parameters, intuitive buttons for navigation and setting values, and a buzzer that provides audible feedback to confirm user actions and alert on important system events.